Move merge/merge-sort device patterns into separate files
